+++ include/s2let_types.h
@@ -89,9 +89,9 @@ typedef struct {
     /*!
      * Wavelet parameter which determines the scale factor between
      * consecutive wavelet scales.
-     * \var int B
+     * \var double B
      */
-    int B;
+    double B;
 
     /*!
      * Upper harmonic band-limit. Only flmn with l < L will be stored
